Substantial update of figure code
The two pieces of code (summarize_records and plot_summaries) weren’t
talking to one another, in that the plotting code pulled in data not
generated by the python script or the summarize_records script.  So, I
re-did the code so that they work sequentially.

I also added code so that we can easily re-order the factors in the
plots.